consensus/pbft: initial validator set when starting pbft backend
