+++ eth/backend.go
@@ -29,7 +29,7 @@ import (
 	"github.com/ethereum/go-ethereum/consensus"
 	"github.com/ethereum/go-ethereum/consensus/clique"
 	"github.com/ethereum/go-ethereum/consensus/ethash"
-	pbft "github.com/ethereum/go-ethereum/consensus/pbft/backends/simple"
+	pbftBackend "github.com/ethereum/go-ethereum/consensus/pbft/backends/simple"
 	"github.com/ethereum/go-ethereum/core"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/core/vm"
@@ -214,7 +214,7 @@ func CreateConsensusEngine(ctx *node.ServiceContext, config *Config, chainConfig
 	}
 	// If PBFT is requested, set it up
 	if chainConfig.PBFT != nil {
-		return pbft.NewBackend(chainConfig.ChainId.Uint64(), uint64(0), uint64(0), ctx.EventMux, ctx.NodeKey())
+		return pbftBackend.New(uint64(0), uint64(0), ctx.EventMux, ctx.NodeKey())
 	}
 
 	// Otherwise assume proof-of-work
